ENH: Update to test building of ITKv4 with DCMTK support
DCMTK and ITKv4 are neede to build the Slicer compliant DWIConvert program.
This patch is the first step towards building DWIConvert as part of DTIPrep.

In addtion, some of the External_proj.cmake files were updated to make
them compatible and compliant with the Slicer build system.  This will
eventually make maintenance cleaner and easier in teh future across
multiple packages.

git-svn-id: https://www.nitrc.org/svn/dtiprep/trunk@212 3db2834d-1276-0410-a171-f2f48b2634cc