BUG: Path to executables was found only after reconfiguring the project
When configuring the project a first time, the path to the executables used for the installation and the packaging were not set properly. CMake was trying to fin the executable but those were not yet built. One had to compile the code first, and then configure the project again to set the paths properly



git-svn-id: https://www.nitrc.org/svn/dtiprep/trunk@244 3db2834d-1276-0410-a171-f2f48b2634cc