STYLE: Useless ITK deprecated define removed.
The ITK_EXPORT define was set to nothing and had no
known remaining purpose.  It was removed to make the
over all code easier to understand.  There was,
understandably, a bit of confusion about the
need for this being pervasive in the code.

It is currently backwards compatible to have
this in code, but at some future point it will
be removed.

git-svn-id: https://www.nitrc.org/svn/dtiprep/trunk@245 3db2834d-1276-0410-a171-f2f48b2634cc